// src/common/enums/user-role.enums.ts
/**
 * Enum representing user roles.
 */
export enum UserRoles {
  /**
   * Represents a regular user.
   */
  USER = 'user',

  /**
   * Represents an administrator.
   */
  ADMIN = 'admin',

  /**
   * Represents a super administrator.
   */
  SUPERADMIN = 'superadmin',

  /**
   * Represents a tutor.
   */
  TUTOR = 'tutor',
}